+++ src/main.ts
@@ -4,6 +4,11 @@ import {
 import { createApp } from 'vue'
 import App from './App.vue'
 
+const APP_ROOT_ID = 'siyuan-anki-linker-app'
+const APP_ROOT_CLASS = 'siyuan-anki-linker-app'
+
+
+
 let plugin: Plugin | null = null
 export function usePlugin(pluginProps?: Plugin): Plugin {
   if (pluginProps) {
@@ -22,8 +27,10 @@ export function init(pluginProps: Plugin) {
   usePlugin(pluginProps)
 
   const div = document.createElement('div')
-  div.classList.add('anki-linker-app', 'fn__none')
-  div.id = 'anki-linker-app'
+  div.classList.add(APP_ROOT_CLASS, 'fn__none')
+  div.id = APP_ROOT_ID
+
+
   app = createApp(App)
   app.mount(div)
   document.body.appendChild(div)